requirements.txt
# Core
numpy>=1.24
scipy>=1.11
pandas>=2.0
pyyaml>=6.0

# Mesh & Medical Imaging
vtk>=9.2
trimesh>=4.0
nibabel>=5.0
scikit-image>=0.21

# Machine Learning
scikit-learn>=1.3
torch>=2.0
torch-geometric>=2.3

# Visualization
matplotlib>=3.7

# Utilities
tqdm>=4.65
